Refactored PresentationSlides component and moved loadGoogleApi file to tools.
