+++ tools/libs/services/loadGoogleApi.js
