+++ tools/outputs/PresentationGenerator/PresentationSlides/PresentationSlides.jsx
@@ -1,9 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
 import ArrowForwardIcon from '@mui/icons-material/ArrowForward';
-import CloudUploadIcon from '@mui/icons-material/CloudUpload';
 import {
+  Alert,
   Box,
   Button,
   Fade,
@@ -13,13 +13,13 @@ import {
   ListItem,
   ListItemButton,
   ListItemText,
-  Typography,
   Snackbar,
-  Alert,
-  CircularProgress,
+  Typography,
 } from '@mui/material';
 
 import styles from './styles';
+
+import GoogleSlidesButton from '@/tools/components/GoogleSlidesButton/GoogleSlidesButton';
 
 /**
  * PresentationSlides component renders the actual presentation view
@@ -28,66 +28,14 @@ import styles from './styles';
  */
 const PresentationSlides = ({ slides, onBackToOutliner }) => {
   const [currentSlideIndex, setCurrentSlideIndex] = useState(0);
-  const [isExporting, setIsExporting] = useState(false);
-  const [notification, setNotification] = useState({ open: false, message: '', severity: 'info' });
-  const [gapiLoaded, setGapiLoaded] = useState(false);
-  const [gisLoaded, setGisLoaded] = useState(false);
-  const [tokenClient, setTokenClient] = useState(null);
-  
+  const [notification, setNotification] = useState({
+    open: false,
+    message: '',
+    severity: 'info',
+  });
+
   // Current slide
   const currentSlide = slides[currentSlideIndex] || { title: '', content: '' };
-
-  // Load the Google API and Identity Services libraries
-  useEffect(() => {
-    // Function to load Google API
-    const loadGapiAndGis = () => {
-      // Load gapi script
-      const gapiScript = document.createElement('script');
-      gapiScript.src = 'https://apis.google.com/js/api.js';
-      gapiScript.async = true;
-      gapiScript.defer = true;
-      gapiScript.onload = () => {
-        initializeGapiClient();
-      };
-      document.body.appendChild(gapiScript);
-
-      // Load Google Identity Services script
-      const gisScript = document.createElement('script');
-      gisScript.src = 'https://accounts.google.com/gsi/client';
-      gisScript.async = true;
-      gisScript.defer = true;
-      gisScript.onload = () => {
-        setGisLoaded(true);
-      };
-      document.body.appendChild(gisScript);
-    };
-
-    loadGapiAndGis();
-  }, []);
-
-  // Initialize GAPI client
-  const initializeGapiClient = async () => {
-    await window.gapi.load('client', async () => {
-      await window.gapi.client.init({
-        apiKey: process.env.NEXT_PUBLIC_GOOGLE_API_KEY,
-        discoveryDocs: ['https://slides.googleapis.com/$discovery/rest?version=v1'],
-      });
-      setGapiLoaded(true);
-    });
-  };
-
-  // Initialize tokenClient when both libraries are loaded
-  useEffect(() => {
-    if (gapiLoaded && gisLoaded) {
-      setTokenClient(
-        window.google.accounts.oauth2.initTokenClient({
-          client_id: process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID,
-          scope: 'https://www.googleapis.com/auth/presentations https://www.googleapis.com/auth/drive.file',
-          callback: '', // Will be set later in the exportToGoogleSlides function
-        })
-      );
-    }
-  }, [gapiLoaded, gisLoaded]);
 
   // Navigate to previous slide
   const handlePrevSlide = () => {
@@ -111,126 +59,6 @@ const PresentationSlides = ({ slides, onBackToOutliner }) => {
   // Close notification
   const handleCloseNotification = () => {
     setNotification({ ...notification, open: false });
-  };
-
-  // Google Slides export function
-  const exportToGoogleSlides = async () => {
-    try {
-      setIsExporting(true);
-      
-      // Check if API libraries are loaded
-      if (!gapiLoaded || !gisLoaded || !tokenClient) {
-        throw new Error('Google API libraries are not fully loaded yet. Please try again.');
-      }
-      
-      // Check if environment variables are available
-      const googleApiKey = process.env.NEXT_PUBLIC_GOOGLE_API_KEY;
-      const googleClientId = process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID;
-      
-      if (!googleApiKey || !googleClientId) {
-        throw new Error('Google API credentials are missing. Please check your environment variables.');
-      }
-
-      // Request an access token
-      tokenClient.callback = async (response) => {
-        if (response.error) {
-          throw new Error(response.error);
-        }
-
-        try {
-          // Create a new presentation
-          const presentationResponse = await window.gapi.client.slides.presentations.create({
-            title: 'Exported Presentation',
-          });
-          
-          const presentationId = presentationResponse.result.presentationId;
-          
-          // Prepare slide creation requests
-          const requests = slides.map((slide, index) => {
-            return {
-              createSlide: {
-                objectId: `slide_${index}`,
-                insertionIndex: index,
-                slideLayoutReference: {
-                  predefinedLayout: 'TITLE_AND_BODY',
-                },
-                placeholderIdMappings: [
-                  {
-                    layoutPlaceholder: {
-                      type: 'TITLE',
-                    },
-                    objectId: `title_${index}`,
-                  },
-                  {
-                    layoutPlaceholder: {
-                      type: 'BODY',
-                    },
-                    objectId: `body_${index}`,
-                  },
-                ],
-              },
-            };
-          });
-          
-          // Create all slides
-          await window.gapi.client.slides.presentations.batchUpdate({
-            presentationId: presentationId,
-            requests: requests,
-          });
-          
-          // Add content to slides
-          const contentRequests = slides.flatMap((slide, index) => [
-            {
-              insertText: {
-                objectId: `title_${index}`,
-                text: slide.title,
-              },
-            },
-            {
-              insertText: {
-                objectId: `body_${index}`,
-                text: slide.content,
-              },
-            },
-          ]);
-          
-          await window.gapi.client.slides.presentations.batchUpdate({
-            presentationId: presentationId,
-            requests: contentRequests,
-          });
-          
-          // Open the created presentation in a new tab
-          window.open(`https://docs.google.com/presentation/d/${presentationId}/edit`, '_blank');
-          
-          setNotification({
-            open: true,
-            message: 'Successfully exported to Google Slides!',
-            severity: 'success',
-          });
-        } catch (error) {
-          console.error('Error creating presentation:', error);
-          setNotification({
-            open: true,
-            message: `Failed to create presentation: ${error.message}`,
-            severity: 'error',
-          });
-        } finally {
-          setIsExporting(false);
-        }
-      };
-
-      // Request an access token with immediate callback
-      tokenClient.requestAccessToken({ prompt: '' });
-      
-    } catch (error) {
-      console.error('Error exporting to Google Slides:', error);
-      setNotification({
-        open: true,
-        message: `Failed to export: ${error.message}`,
-        severity: 'error',
-      });
-      setIsExporting(false);
-    }
   };
 
   return (
@@ -284,21 +112,10 @@ const PresentationSlides = ({ slides, onBackToOutliner }) => {
             >
               Back to Outliner
             </Button>
-            
-            <Button
-              startIcon={isExporting ? <CircularProgress size={16} color="inherit" /> : <CloudUploadIcon />}
-              onClick={exportToGoogleSlides}
-              disabled={isExporting || slides.length === 0 || !gapiLoaded || !gisLoaded}
-              sx={{
-                color: styles.slideTitleProps?.color || '#AC92FF',
-                backgroundColor: '#2A1B4A',
-                '&:hover': { backgroundColor: '#3A2B5A' },
-                borderRadius: '6px',
-                padding: '6px 16px',
-              }}
-            >
-              {isExporting ? 'Exporting...' : !gapiLoaded || !gisLoaded ? 'Loading...' : 'Export to Google Slides'}
-            </Button>
+            <GoogleSlidesButton
+              slides={slides}
+              setNotification={setNotification}
+            />
           </Grid>
 
           {/* Main content with sidebar and slide view */}
@@ -353,11 +170,12 @@ const PresentationSlides = ({ slides, onBackToOutliner }) => {
                           <Typography
                             sx={{
                               color: styles.slideTitleProps?.color || '#AC92FF',
-                              fontWeight: currentSlideIndex === index ? 'bold' : 'normal',
+                              fontWeight:
+                                currentSlideIndex === index ? 'bold' : 'normal',
                               width: '100%',
                               overflow: 'hidden',
                               textOverflow: 'ellipsis',
-                              whiteSpace: 'nowrap'
+                              whiteSpace: 'nowrap',
                             }}
                           >
                             {`${index + 1}. ${slide.title}`}
@@ -372,10 +190,13 @@ const PresentationSlides = ({ slides, onBackToOutliner }) => {
                               textOverflow: 'ellipsis',
                               whiteSpace: 'nowrap',
                               width: '200px', // Fixed width
-                              maxWidth: '100%'
+                              maxWidth: '100%',
                             }}
                           >
-                            {slide.content.split(' ').slice(0, 6).join(' ') + '...'}
+                            {`${slide.content
+                              .split(' ')
+                              .slice(0, 6)
+                              .join(' ')}...`}
                           </Typography>
                         }
                       />
@@ -415,7 +236,7 @@ const PresentationSlides = ({ slides, onBackToOutliner }) => {
                   justifyContent: 'flex-start', // Changed from center
                   textAlign: 'center',
                   mb: 4,
-                  ml:-2,
+                  ml: -2,
                   width: '100%',
                   border: 'none',
                   overflow: 'hidden', // Hide overflow
@@ -491,16 +312,16 @@ const PresentationSlides = ({ slides, onBackToOutliner }) => {
             </Grid>
           </Grid>
         </Grid>
-        
+
         {/* Notification for export status */}
-        <Snackbar 
-          open={notification.open} 
-          autoHideDuration={6000} 
+        <Snackbar
+          open={notification.open}
+          autoHideDuration={6000}
           onClose={handleCloseNotification}
           anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
         >
-          <Alert 
-            onClose={handleCloseNotification} 
+          <Alert
+            onClose={handleCloseNotification}
             severity={notification.severity}
             sx={{ width: '100%' }}
           >
